fix the error

+++ src/actumcrypto/actumcrypto.h
@@ -7,6 +7,6 @@
 #include "rpc/rpcwallet.h"
 #include "script/standard.h"
 
-void RewardMinedBlock(CWallet* pwallet, int amount);
+void RewardMinedBlock(CWallet* pwallet, double amount);
 
 #endif // ACTUMCRYPTO_H
